% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weight handling functions.R
\name{as.w8margin.data.frame}
\alias{as.w8margin.data.frame}
\title{Convert Data Frame to w8margin Object}
\usage{
as.w8margin.data.frame(target, varname = NULL, levels = NULL, 
 samplesize = NULL, rebase.tol = .01).
}
\arguments{
\item{target}{Data frame for conversion to a \code{w8margin} object. Must
have one numeric column specifying target values foreach level, and either
a non-numeric column or row names specifying the name of each target level.}

\item{varname}{Character vector specifying the name of the observed variable
that the \code{w8margin} object should match. Optional for two-column data
frames; if omitted, If NULL, default value is the column name of the
non-numeric column).}

\item{levels}{Character vector of length \code{nrow(target)} to
override default target levels of w8margin.}

\item{samplesize}{Number with the desired target sample size for the w8margin
object. Defaults to \code{sum(target$Freq)}.}

\item{rebase.tol}{Numeric betweeen 0 and 1. Generates a warning if
targets are rebased, and the rebased sample sizes differs by more than this
percentage.}
}
\value{
An object of class w8margin, with specified varname and samplesize.
}
\description{
Takes a data frame (with one or two columns, see "target" param
  below). Converts to a \code{w8margin} object with specified name and sample
  size (rebasing if necessary).
}
