% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weight handling functions.R
\name{as.w8margin.matrix}
\alias{as.w8margin.matrix}
\title{Convert Matrix to w8margin Object}
\usage{
as.w8margin.matrix(target, varname, levels = NULL, samplesize = NULL,
  byrow = TRUE, rebase.tol = .01)
}
\arguments{
\item{target}{Matrix with row and column names, for conversion to a w8margin
object. By default (unless overridden by levels), the target levels
of \code{w8margin} will come from the interaction of row and column names.}

\item{varname}{Character vector specifying the name of the observed variable
that the w8margin object should match.}

\item{levels}{Character vector of length \code{ncol(target) * nrow(target)} to override
default target levels of w8margin.}

\item{samplesize}{Numeric with the desired target sample size for the
w8margin object. Defaults to \code{sum(targe)}.}

\item{byrow}{Logical. If FALSE, the elements within columns will be adjacent
in the resulting w8margin object, otherwise elements within rows will be
adjacent.}

\item{rebase.tol}{Numeric betweeen 0 and 1. If targets are rebased, and
the rebased sample sizes differs from the original sample size by more than
this percentage, generates a warning.}
}
\value{
An object of class w8margin, with specified varname and samplesize.
}
\description{
Takes a matrix (with row and column names), and converts to a
  \code{w8margin} object with specified name and sample size (rebasing if
  necessary).
}
