% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weight handling functions.R
\name{as.w8margin.numeric}
\alias{as.w8margin.numeric}
\title{Convert Numeric Vector to w8margin Object}
\usage{
as.w8margin.numeric(target, varname, levels = NULL, samplesize =
  NULL, rebase.tol = .01)
}
\arguments{
\item{target}{Named vector, for conversion to a w8margin object. By default
(unless overridden by \code{levels}), the target levels of w8margin will
come from the names of the vector.}

\item{varname}{Character vector specifying the name of the observed variable
 that the w8margin object should match.
@param levels Character vector of length \code{length(target)} to override default
 target levels of w8margin.}

\item{samplesize}{Integer with the desired target sample size for the
w8margin object. Defaults to \code{sum(target)}.}

\item{rebase.tol}{Numeric betweeen 0 and 1. If targets are rebased, and
the rebased sample sizes differs from the original sample size by more than
this percentage, generates a warning.}
}
\value{
An object of class w8margin, with specified varname and samplesize.
}
\description{
Takes a named numeric vector, and converts to a \code{w8margin} object
  with specified variable name and sample size (rebasing if necessary).
}
