% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weight handling functions.R
\name{as.w8margin}
\alias{as.w8margin}
\alias{w8margin}
\alias{as.w8margin.matrix}
\alias{as.w8margin.data.frame}
\alias{as.w8margin.numeric}
\title{Weight Margin Objects}
\usage{
as.w8margin(target, varname, levels = NULL, samplesize = NULL,
  rebase.tol = .01)

\method{as.w8margin}{matrix}(target, varname, levels = NULL,
  samplesize = NULL, byrow = TRUE, rebase.tol = 0.01)

\method{as.w8margin}{data.frame}(target, varname = NULL, levels = NULL,
  samplesize = NULL, rebase.tol = 0.01)

\method{as.w8margin}{numeric}(target, varname, levels = NULL,
  samplesize = NULL, rebase.tol = 0.01)
}
\arguments{
\item{target}{Numbers specifying the desired target distribution of a
categorical variable, after rake weighting. Can be a numeric vector,
numeric matrix, or data frame with one (and only one) numeric column.
Unless \code{levels} is specified, vectors and matrices must be named, and
data frames must have a character or factor column specifying names. See
details.}

\item{varname}{Character vector specifying the name of the observed variable
that the \code{w8margin} object should match.}

\item{levels}{Optional character vector, specifying which numeric elements of
\code{target} match with each factor level in the observed data. Overrides
names specified in \code{target}.}

\item{samplesize}{Numeric with the desired target sample size for the
w8margin object. Defaults to the sum of \code{target}.}

\item{byrow}{Logical, specifying how matrix targets should be converted to vectors. 
If FALSE, the elements within columns will be adjacent
in the resulting w8margin object, otherwise elements within rows will be
adjacent.}

\item{rebase.tol}{Numeric betweeen 0 and 1. If targets are rebased, and the
rebased sample sizes differs from the original sample size by more than
this percentage, generates a warning.}
}
\value{
An object of class w8margin, with specified variable name and sample size.
}
\description{
Creates an object of class \code{w8margin}, representing the
  desired target distribution of a single categorical variable after
  weighting. Used as input for \code{\link{rakesvy}}, \code{\link{rakew8}},
  \code{\link[survey]{rake}}, and \code{\link[survey]{postStratify}}
  functions. Methods exist for numeric vectors, matrices, and data frames
  (see details).
}
\details{
w8margin objects are inputs to the \code{\link{rakesvy}},
  \code{\link{rakew8}}, \code{\link[survey]{rake}}, and
  \code{\link[survey]{postStratify}} functions. These functions require a
  specific, highly-structured input format. For flexibility,
  \code{as.w8margin} can be used to convert a variety of common inputs into
  the format needed by these functions.

\code{as.w8margin} has methods for numeric vectors, numeric matrices, and
  data frames. Each method has multiple ways of determining how to match
  numeric elements of \code{target} with factor levels in the observed data.
  For numeric vector and matrix inputs, the default is to match based on the
  name of each element (for vectors) or the interaction of row and column
  names of each element (for matrices). These names can be overridden by
  specifying the \code{levels} parameter.

Data frame inputs must have either one or two columns. Two-column
  data frames must have one numeric column and one character column. The
  numeric column specifies the target distribution, while the character
  column specifies how to match numeric elements with factor levels in the
  observed data. If \code{varname} is NULL, a default value will be taken
  from the name of the non-numeric column.

One-column data frames must have a numeric column. Row names are
  converted to a charater column in order to match numeric elements with
  factor levels in the observed data. One-column data frames must specify a
  \code{varname} parameter, and (unless \code{levels} is specified) must have
  non-default row names. The \code{levels} parameter can be used with both
one- and two-column data frames.

Technically, \code{w8target} objects are data frames with two
  columns. The first column specifies levels in the observed factor variable,
  and the *name* of the first column indicates the name of the observed
  factor variable. The second column is named "Freq" and indicates the
  desired post-raking frequency of each category (as a *count* rather than percentage). 
  The structure is designed for compatibility with the srvy package.
  Because frequency is specified as a count, \code{\link{rakesvy}} and \code{\link{rakew8}} 
  re-call \code{as.w8margin} whenever weighting a data set to a new observed sample size. 
  Weight margins must be manually re-calculated for new sample sizes when using
  \code{\link[survey]{postStratify}} or \code{\link[survey]{rake}}.
}
