% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weight handling functions.R
\name{as.w8target.data.frame}
\alias{as.w8target.data.frame}
\title{Convert Data Frame to w8target Object}
\usage{
as.w8target.data.frame(target, varname = NULL, samplesize = NULL,
  forcedLevels = NULL, rebaseTolerance = .01).
}
\arguments{
\item{target}{Data frame for conversion to a \code{w8target} object. Must have two
variables: one numeric variable named "Freq" specifying target numbers, and
one variable specifying names for each target level (unless overridden by
forcedLevels).}

\item{varname}{Character vector specifying the name of the observed variable
that the \code{w8target} object should match. If NULL, default value is the name
of the data frame variable other than "Freq")}

\item{samplesize}{Integer with the desired target sample size for the
w8target object. Defaults to \code{sum(target$Freq)}.}

\item{forcedLevels}{Character vector of length \code{nrow(target)} to override default
target levels of w8target.}

\item{rebaseTolerance}{Numeric betweeen 0 and 1. Generates a warning if targets are rebased, and
the rebased sample sizes differs by more than this percentage.,}
}
\value{
An object of class w8target, with specified varname and samplesize.
}
\description{
Takes a data frame with two columns (one named "Freq" specifying
  target levels names, and one specifying the target number for each level).
  Converts to a \code{w8target} object with specified name and sample size
  (rebasing if necessary).
}
