% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weight handling functions.R
\name{as.w8target.matrix}
\alias{as.w8target.matrix}
\title{Convert Matrix to w8target Object}
\usage{
as.w8target.matrix(target, varname, samplesize = NULL, forcedLevels =
  NULL, byrow = TRUE, rebaseTolerance = .01)
}
\arguments{
\item{target}{Matrix with row and column names, for converstion to a w8target
object. By default (unless overridden by forcedLevels), the target levels
of \code{w8target} will come from the interaction of row and column names.}

\item{varname}{Character vector specifying the name of the observed variable
that the w8target object should match}

\item{samplesize}{Integer with the desired target sample size for the
w8target object. Defaults to \code{sum(targe)}.}

\item{forcedLevels}{Character vector of length \code{ncol(target) * nrow(target)} to override
default target levels of w8target}

\item{byrow}{Logical. If FALSE, the elements within columns will be adjacent
in the resulting w8target object, otherwise elements within rows will be
adjacent.}

\item{rebaseTolerance}{Numeric betweeen 0 and 1. If targets are rebased, and
the rebased sample sizes differs from the original sample size by more than
this percentage, generates a warning.}
}
\value{
An object of class w8target, with specified varname and samplesize.
}
\description{
Takes a matrix (with row and column names), and converts to a
  \code{w8target} object with specified name and sample size (rebasing if
  necessary).
}
