% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weight handling functions.R
\name{as.w8target.numeric}
\alias{as.w8target.numeric}
\title{Convert Numeric Vector to w8target Object}
\usage{
as.w8target.numeric(target, varname, samplesize = NULL, forcedLevels =
  NULL, rebaseTolerance = .01)
}
\arguments{
\item{target}{Named vector, for converstion to a w8target object. By default
(unless overridden by \code{forcedLevels}), the target levels of w8target will
come from the names of the vector.}

\item{varname}{Character vector specifying the name of the observed variable
that the w8target object should match.}

\item{samplesize}{Integer with the desired target sample size for the
w8target object. Defaults to \code{sum(target)}.}

\item{forcedLevels}{Character vector of length \code{length(target)} to override default
target levels of w8target.}

\item{rebaseTolerance}{Numeric betweeen 0 and 1. If targets are rebased, and
the rebased sample sizes differs from the original sample size by more than
this percentage, generates a warning.}
}
\value{
An object of class w8target, with specified varname and samplesize.
}
\description{
Takes a named numeric vector, and converts to a \code{w8target} object
  with specified variable name and sample size (rebasing if necessary).
}
