% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weight handling functions.R
\name{checkTargetMatch}
\alias{checkTargetMatch}
\title{Check Whether w8target Object Matches Observed Variable}
\usage{
checkTargetMatch(w8target, observedVar, exact = FALSE, refactor =
  FALSE)
}
\arguments{
\item{w8target}{w8target object (or data frame that in the format specified
by rake, which can behave as a w8target object).}

\item{observedVar}{factor variable (or, if refactor = FALSE, a variable that
can be coerced to factor).}

\item{exactMatch}{logical, specifying if levels in w8target must be in the
same order as factor levels in observedVar.}

\item{refactor}{logial, specifying whether to factor variable before checking
match.}
}
\value{
A logical, indicating whether w8target is compatible with observedVar.
}
\description{
Checks whether specified \code{w8target} object and observedVar are
  compatible, and are expected to produce valid call to rake. Returns a
  logical true/false, and generates warning messages to specify likely issue.
  Intended to help quickly diagnose incompatibilities between w8targets and
  observed data.
}
