% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w8margin_functions.R
\name{eff.n}
\alias{eff.n}
\title{Kish's Effective Sample Size}
\usage{
\code{eff.n(design)}
}
\arguments{
\item{design}{An \code{\link[survey]{svydesign}} object, presumably with
design or post-stratification weights.}
}
\description{
Computes Kish's effective sample size for a
  \code{\link[survey]{svydesign}} object. Computed using the formula
  \code{sum(weights(design)) ^ 2 / sum(weights(design) ^ 2)}.
}
\details{
Kish's effective sample size is a frequently-used, general metric to
  indicate how much uncertainty and error increase due to weighting. However,
  it is less valid than the standard errors produced by
  \code{\link[survey]{svymean}} and related functions from the {survey}
  package: it ignores clustering and stratification in sample designs, and
  covariance between weighting variables and outcome variables.

The *weighting efficiency* of a sample is a closely related metric,
  which shares all the problems of effective sample size. It can be
  calculated as \code{eff.n(design) / sum(weights(design))}.
}
