% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_functions.R
\name{eff_n}
\alias{eff_n}
\alias{eff_n.survey.design}
\alias{eff_n.numeric}
\alias{weight_eff}
\alias{weight_eff.survey.design}
\alias{weight_eff.numeric}
\title{Effective Sample Size and Weighting Efficiency}
\usage{
eff_n(x)

\method{eff_n}{survey.design}(x)

\method{eff_n}{numeric}(x)

weight_eff(x)

\method{weight_eff}{survey.design}(x)

\method{weight_eff}{numeric}(x)
}
\arguments{
\item{x}{Either an object from \code{\link[survey]{svydesign}} (with
weights), or a vector of case weights.}
}
\description{
Computes Kish's effective sample size or weighting efficiency for a
  \code{\link[survey]{svydesign}} object.
}
\details{
Kish's effective sample size is a frequently-used, general metric to
  indicate how much uncertainty and error increase due to weighting. 
  Effective sample size is calculated as \code{sum(weights(design))^2 / sum(weights(design)^2)}. 
  Weighting efficiency is \code{eff_n(design) / sum(weights(design))}.

While weighting efficiency and effective sample size are frequently used,
 they are less valid than the standard errors produced by
  \code{\link[survey]{svymean}} and related functions from the {survey}
  package. In particular, they ignore clustering and stratification in 
  sample designs, and covariance between weighting variables and outcome variables.
  As such, these metrics should be used with caution
}
\examples{
gles17_weighted <- rakesvy(design = gles17, 
    gender ~ c("Male" = .495, "Female" = .505),
    eastwest ~ c("East Germany" = .195, "West Germany" = .805)
)

eff_n(gles17_weighted)
weight_eff(gles17_weighted)
}
