% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weight handling functions.R
\name{rakesvy}
\alias{rakesvy}
\alias{rakew8}
\title{Flexibly Calculate Rake Weights}
\usage{
rakesvy(design, weightTargets, samplesize = "fromData", 
  matchLevelsBy = "name", matchVarsBy = "listname", rebaseTolerance = .01, ...)

rakew8(design, weightTargets, samplesize = "fromData",
  matchLevelsBy = "name", matchVarsBy = "listname",
  rebaseTolerance = 0.01, ...)
}
\arguments{
\item{design}{An \code{\link[survey]{svydesign}} object, or a data frame that
can be coerced to an svydesign object (assuming no clustering or design
weighting)}

\item{weightTargets}{A list of w8target objects, or other objects that can be
coerced to weightTargets.}

\item{samplesize}{Either an integer specifying the desired post-raking sample
size, or a character string "fromData" or "fromTargets" specfiying how to
calculate the desired sample size (see details).}

\item{matchLevelsBy}{A character string that specifies how to match levels in
the target with the observed data, either "name" (the default) or "order"
(see details).}

\item{matchVarsBy}{A character  string that specificies how elements of
weightTargets are matched with variables in design, either "listname" (the
default) or "colname" (see details).}

\item{rebaseTolerance}{Numeric betweeen 0 and 1. If targets are rebased, and
the rebased sample sizes differs from the original sample size by more than
this percentage, generates a warning.}
}
\value{
rakesvy resutns an \code{svydesign} object with rake weights applied. Any changes
  made the variables in \code{design} in order to call \code{rake}, such as
  dropping empty factor levels, are temporary and \emph{not} returned in the
  output object.

rakew8 returns a vector of eights. This avoids creating
  duplicated svydesign objects, which can be useful when calculating multiple
  sets of weights for the same data.
}
\description{
Calculates rake weights, using flexible flexible syntax for data
  and weight target specification. Runs pre-processing, then calls
  \code{\link[survey]{rake}} to comptue weights. rakesvy returns a
  weighted \code{svydesign} object, while rakew8 returns a vector of
  weights.
}
\details{
rakesvy and rakew8 are a wrapper for \code{\link[survey]{rake}} that wrangles
  observed data and targets. It cleans matches weight targets to observed
  variables, cleans both targets and observed varaibles, and then checks the
  validity of weight targets (partially by calling
  \code{\link{checkTargetMatch}} before raking. It also allows a weight
  target of zero, assigns an automatic weight of zero to cases on this target
  level.

Weight target levels can be matched with observed variable levels in
  two ways, specified via the \code{matchLevelsBy} parameter. "name" (the
  default) matches based on name, disregarding order (so a "male" level in
  the weight target will be matched with a "male" level in the observed
  data). "order" matches based on order, disregarding name (so the first
  level or row of the target will match with the first level of the observed
  factor variable).

Weight targets can also be matched to observed variables in two
  ways, specified via the \code{matchVarsBy} paramter. The default,
  "listname", indicated that the names of elements in the list weightTargets
  should indicate variables in the design object. The alternative, "colname"
  specifies that the non-"Freq" column name of each item in the list
  weightTargets should indicate a matching variable in the design object;
  this will only work for weight targets in a \code{w8target} or
  \code{data.frame} format.

The desired sample size (in other words, the desired sum of weights
  after raking)  is specifeid via the \code{samplesize} parameter. This can
  be a numeric value. Alternatively, "fromData" specifies that the observed
  sample size before weighting (taken from \code{sum(weights(design))} if
  appliable, or \code{nrow} if not); "fromTargets" specifies that the total
  sample sizes in target objects should be followed, and should only be used
  if all targets specify the same sample size.
}
