% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weight handling functions.R
\name{target.matched}
\alias{target.matched}
\title{Check Whether w8margin Object Matches Observed Variable}
\usage{
target.matched(w8margin, observed, exact = FALSE, refactor =
  FALSE)
}
\arguments{
\item{w8margin}{w8margin object (or data frame that in the format specified
by rake, which can behave as a w8margin object).}

\item{observed}{factor variable (or, if refactor = FALSE, a variable that
can be coerced to factor).}

\item{refactor}{logial, specifying whether to factor variable before checking
match.}
}
\value{
A logical, indicating whether w8margin is compatible with observed.
}
\description{
Checks whether specified \code{w8margin} object and observed are
  compatible, and are expected to produce valid call to rake. Returns a
  logical true/false, and generates warning messages to specify likely issue.
  Intended to help quickly diagnose incompatibilities between w8margins and
  observed data.
}
