% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w8margin_functions.R
\name{w8margin.matched}
\alias{w8margin.matched}
\title{Check Whether w8margin Object Matches Observed Variable}
\usage{
w8margin.matched(w8margin, observed, refactor = FALSE)
}
\arguments{
\item{w8margin}{w8margin object, or other object type that can be coerced to
w8margin with a temporary variable name.}

\item{observed}{factor variable (or, if \code{refactor = TRUE}, a variable that can
be coerced to factor).}

\item{refactor}{logical, specifying whether to factor observed variable before checking
match.}
}
\value{
A logical, indicating whether w8margin is compatible with observed.
}
\description{
Checks whether specified \code{\link{w8margin}} object and variable in observed
  data are compatible, and are expected to produce valid call to
  \code{\link[survey]{rake}}. Returns a logical true/false, and generates
  warning messages to specify likely issues. Intended to help quickly
  diagnose incompatibilities between w8margins and observed data.
}
