% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w8margin_functions.R
\name{w8margin_matched}
\alias{w8margin_matched}
\title{Check Whether w8margin Object Matches Observed Variable}
\usage{
w8margin_matched(w8margin, observed, refactor = FALSE)
}
\arguments{
\item{w8margin}{w8margin object, or other object type that can be coerced to
w8margin with a temporary variable name.}

\item{observed}{factor variable (or, if \code{refactor = TRUE}, a variable that can
be coerced to factor).}

\item{refactor}{logical, specifying whether to factor observed variable before checking
match.}
}
\value{
A logical, indicating whether w8margin is compatible with observed.
}
\description{
Checks whether specified \code{\link{w8margin}} object and variable in observed
  data are compatible, and are expected to produce valid call to
  \code{\link[survey]{rake}}. Returns a logical true/false, and generates
  warning messages to specify likely issues. Intended to help quickly
  diagnose incompatibilities between w8margins and observed data.
}
\details{
This function is primarily intended for internal use by \code{\link{rakesvy}}.
  However, it may be useful to call directly, when manually calling \code{\link[survey]{rake}}
  instead of using the Rakehelper interface.
}
\examples{
gender_w8margin <- as.w8margin(
    c(Male = .49, Female = .51),
    varname = "gender", 
    samplesize = 2179)
# Returns TRUE
w8margin_matched(gender_w8margin, gles17$gender)

gender_w8margin_alt <- as.w8margin(
    c(man = .49, woman = .51), 
    varname = "gender", 
    samplesize = 2179)
# Returns FALSE - level names in gles17$gender do not match level names in gender_w8margin_alt
w8margin_matched(gender_w8margin_alt, gles17$gender)

agecat_50plus_w8margin <- as.w8margin(
    c("50-59" = .35, "60-69" = .27, ">=70" = .38),
    varname = "educ",
    samplesize = 2179
)
gles17_50plus <- gles17[gles17$agecat \%in\% c("50-59", "60-69", ">=70"),]
# Returns FALSE - gles17$agecat has empty factor levels for <=29, 30-39, 40-49
w8margin_matched(agecat_50plus_w8margin, gles17_50plus$agecat)
# Returns TRUE - gles17$agecat is refactored to drop empty levels
w8margin_matched(agecat_50plus_w8margin, gles17_50plus$agecat, refactor = TRUE)
}
